feat(subscriptions): add optional web UI for subscription management - configurable in procyon-settings - can optionally be password-protected - set up vite and build assets - add blade-formatter and workflow

+++ app/Console/Commands/UpdateSubscriptions.php
@@ -35,17 +35,19 @@ class UpdateSubscriptions extends Command {
             }
         }
 
-        // Remove any that are on record but absent from the config file
-        // Note that this does not delete old digests!
-        $subscriptions = Subscription::pluck('url')->toArray();
-        $missing = array_merge(array_diff(config('procyon-settings.feeds'), $subscriptions), array_diff($subscriptions, config('procyon-settings.feeds')));
+        if (!config('procyon-settings.web-interface')) {
+            // Remove any that are on record but absent from the config file
+            // Note that this does not delete old digests!
+            $subscriptions = Subscription::pluck('url')->toArray();
+            $missing = array_merge(array_diff(config('procyon-settings.feeds'), $subscriptions), array_diff($subscriptions, config('procyon-settings.feeds')));
 
-        if (count($missing)) {
-            foreach ($missing as $sub) {
-                $subscription = Subscription::where('url', $sub)->first();
-                if ($subscription) {
-                    // $subscription->digests->delete();
-                    $subscription->delete();
+            if (count($missing)) {
+                foreach ($missing as $sub) {
+                    $subscription = Subscription::where('url', $sub)->first();
+                    if ($subscription) {
+                        // $subscription->digests->delete();
+                        $subscription->delete();
+                    }
                 }
             }
         }
